test ordering
